🔁 :: get the content from the API route
