+++ package.json
@@ -33,6 +33,7 @@
 		"lint-staged": "^13.2.0",
 		"postcss": "^8.4.21",
 		"prettier": "^2.8.6",
+		"swr": "^2.2.2",
 		"tailwindcss": "^3.3.3"
 	}
 }
