+++ src/pages/api/events/data.ts
@@ -1,0 +1,39 @@
+import { Event } from './type';
+
+export const events: Event[] = [
+	{
+		name: 'Hacktoberfest 2023',
+		description:
+			'Third edition of the Hacktoberfest, celebrating the open source world.',
+		// month/day/year
+		date: '10/23/2023',
+		type: 'upcoming',
+	},
+	{
+		name: 'Atelier Dev Etudiant CIGA',
+		description:
+			"Une belle journ2e d'apprentissage avec les 2tudiants de CIGA.",
+		date: '05/27/2023',
+		type: 'previous',
+	},
+	{
+		name: 'Hacktoberfest 2023',
+		description:
+			'Third edition of the Hacktoberfest, celebrating the open source world.',
+		date: '10/15/2022',
+		type: 'previous',
+	},
+	{
+		name: 'Hacktoberfest 2021',
+		description:
+			'First edition of the Hacktoberfest, celebrating the open source world.',
+		date: '10/23/2021',
+		type: 'previous',
+	},
+	{
+		name: 'Xarala Tour Dakar 2021',
+		description: 'Cloture de la 1ere phase du Xarala Tour 2021.',
+		date: '07/3/2023',
+		type: 'previous',
+	},
+];
