+++ src/pages/api/events/type.ts
@@ -1,0 +1,6 @@
+export type Event = {
+	name: string;
+	type: 'upcoming' | 'previous';
+	description: string;
+	date: string;
+};
